Generator rewrite to support new postfix syntax
This is a complete rewrite of the generator code, factoring handling for
individual TJSON types into separate Ruby classes.
