+++ lib/tjson/datatype/binary.rb
@@ -7,6 +7,10 @@ module TJSON
 
     # Base16-serialized binary data
     class Binary16 < Binary
+      def tag
+        "b16"
+      end
+
       def convert(str)
         raise TJSON::TypeError, "expected String, got #{str.class}: #{str.inspect}" unless str.is_a?(::String)
         raise TJSON::ParseError, "base16 must be lower case: #{str.inspect}" if str =~ /[A-F]/
@@ -14,10 +18,18 @@ module TJSON
 
         [str].pack("H*")
       end
+
+      def generate(binary)
+        binary.unpack("H*").first
+      end
     end
 
     # Base32-serialized binary data
     class Binary32 < Binary
+      def tag
+        "b32"
+      end
+
       def convert(str)
         raise TJSON::TypeError, "expected String, got #{str.class}: #{str.inspect}" unless str.is_a?(::String)
         raise TJSON::ParseError, "base32 must be lower case: #{str.inspect}" if str =~ /[A-Z]/
@@ -26,10 +38,18 @@ module TJSON
 
         ::Base32.decode(str.upcase).force_encoding(Encoding::BINARY)
       end
+
+      def generate(binary)
+        Base32.encode(binary).downcase.delete("=")
+      end
     end
 
     # Base64-serialized binary data
     class Binary64 < Binary
+      def tag
+        "b64"
+      end
+
       def convert(str)
         raise TJSON::TypeError, "expected String, got #{str.class}: #{str.inspect}" unless str.is_a?(::String)
         raise TJSON::ParseError, "base64url only: #{str.inspect}" if str =~ %r{\+|\/}
@@ -40,6 +60,10 @@ module TJSON
         str = str.ljust((str.length + 3) & ~3, "=") if (str.length % 4).nonzero?
 
         ::Base64.urlsafe_decode64(str)
+      end
+
+      def generate(binary)
+        Base64.urlsafe_encode64(binary).delete("=")
       end
     end
   end
