--- lib/tjson/datatype/scalar.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module TJSON
-  class DataType
-    # Scalar types
-    class Scalar < TJSON::DataType
-      def scalar?
-        true
-      end
-
-      def inspect
-        "#<#{self.class}>"
-      end
-    end
-  end
-end
