+++ spec/tjson/datatype/object_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe TJSON::DataType::Object do
+  describe "#convert" do
+    it "needs tests!"
+  end
+
+  describe "#generate" do
+    it "needs tests!"
+  end
+end
